[refactor] Voucher, VoucherProvider로 이름 변경완료
