+++ src/main/java/com/pintoss/gitftmall/common/exceptions/client/DuplicateVoucherProviderNameException.java
@@ -4,12 +4,12 @@ import com.pintoss.gitftmall.common.exceptions.ErrorCode;
 import lombok.Getter;
 
 @Getter
-public class DuplicateProductNameException extends BadRequestException {
-    public DuplicateProductNameException(String message){
+public class DuplicateVoucherProviderNameException extends BadRequestException {
+    public DuplicateVoucherProviderNameException(String message){
         super(message);
     }
 
-    public DuplicateProductNameException(ErrorCode errorCode, String message){
+    public DuplicateVoucherProviderNameException(ErrorCode errorCode, String message){
         super(errorCode, message);
     }
 }
